# backend/specs/001-book-rag-chatbot/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Book RAG Chatbot Backend API
  description: API for a Retrieval-Augmented Generation (RAG) chatbot that allows users to ask questions about book content
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://your-deployment-url.com
    description: Production server

paths:
  /query-global:
    post:
      summary: Query the full book content
      description: Submit a question about the book content and receive an answer based on the entire book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalQueryRequest'
      responses:
        '200':
          description: Successful response with the answer and retrieved chunks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Bad request - invalid input
        '500':
          description: Internal server error
      tags:
        - Query

  /query-selection:
    post:
      summary: Query only user-selected text
      description: Submit a question with selected text and receive an answer based only on the provided text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectionQueryRequest'
      responses:
        '200':
          description: Successful response with the answer and retrieved chunks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Bad request - invalid input
        '500':
          description: Internal server error
      tags:
        - Query

  /status:
    get:
      summary: Health check
      description: Check if the service is running and healthy
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
        '500':
          description: Service is not healthy
      tags:
        - Health

components:
  schemas:
    GlobalQueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The user's question about the book content
          example: "What are the main themes in this book?"
        conversation_id:
          type: string
          description: Optional conversation identifier for maintaining context
          example: "conv_12345"

    SelectionQueryRequest:
      type: object
      required:
        - query
        - selected_text
      properties:
        query:
          type: string
          description: The user's question about the selected text
          example: "What does this passage mean?"
        selected_text:
          type: string
          description: The text selected by the user
          example: "This is the selected text that the user wants to ask about..."
        conversation_id:
          type: string
          description: Optional conversation identifier for maintaining context
          example: "conv_12345"

    QueryResponse:
      type: object
      required:
        - answer
        - retrieved_chunks
        - query_id
      properties:
        answer:
          type: string
          description: The generated answer based on the book content
          example: "The main themes in this book include..."
        retrieved_chunks:
          type: array
          items:
            $ref: '#/components/schemas/RetrievedChunk'
          description: List of chunks that were retrieved and used to generate the answer
        query_id:
          type: string
          description: Unique identifier for the query
          example: "query_67890"
        conversation_id:
          type: string
          description: Conversation identifier if context was maintained
          example: "conv_12345"

    RetrievedChunk:
      type: object
      required:
        - chunk_id
        - content
        - relevance_score
      properties:
        chunk_id:
          type: string
          description: Unique identifier for the retrieved chunk
          example: "chunk_abc123"
        content:
          type: string
          description: The content of the retrieved chunk
          example: "This is the content of the book chunk that was retrieved..."
        relevance_score:
          type: number
          format: float
          description: Relevance score of the chunk to the query
          example: 0.85

tags:
  - name: Query
    description: Endpoints for querying the book content
  - name: Health
    description: Health check endpoints